Update form service url
Due to an update in the form service that effects the connection of this touchpoint an update to the env variabels has been made to match the current url of form service. Along with this the naming of the env variabel has been changed to follow the naming conventions of other env variabels.