Refactor/forms endpoint (#26)
* Handling connection issues from upstream services.

The touchpoint now throws an error 502 (BAD REQUEST) if it cannot connect to the form service. This implementation can be further refactored to be used on other places aswell

* Fixed endpoint for get questions.

The endpoint for retirving questions in the form component was inconsistent. Instead of targeting the forms/:id/questions route at the service, the /forms/:id endpoint was the target.

* Update error.js